Skip /etc/zprofile removal if already deleted.

+++ commands/install/zsh.go
@@ -11,7 +11,9 @@ func Zsh() {
 
   // delete /etc/zprofile - added by os x 10.11
   // path_helper conflicts - http://www.zsh.org/mla/users/2015/msg00727.html
-  util.Run("sudo rm -f /etc/zprofile")
+  if util.IsFileExists("/etc/zprofile") {
+    util.Run("sudo rm -f /etc/zprofile")
+  }
 
   // ensure antibody is installed
   if util.IsCommand("brew") && !util.IsCommand("antibody") {
